requirements.txt
anthropic
openai
supabase
python-dotenv